refactor(test): extract fake Express middleware helpers to shared util
Both test/server/test_csrfGuard.ts and the new
test_csrfGuard_env_wiring.ts (#1463) carried duplicate copies of
FakeReq / FakeRes interfaces plus makeReq / makeReqWithRawOrigin /
makeRes factories. Move them to test/server/helpers/fakeExpressMiddleware.ts
and import from both sites — no behaviour change, net -69 lines.
